hubspot_mcp_lambda/requirements.txt
bedrock-agentcore>=0.1.0
hubspot-api-client>=12.0.0
